409760 "You've been warned:" zero-tolerance for warnings in mozilla/camino.  I got asked why I gave such a few warnings.  There are no compiler warnings in mozilla/camino at all, and we now build everything with -Wall -Werror.  r=smorgan r=ardissone a/1.6b1=me

+++ camino/PreferencePanes/Tabs/Tabs.mm
@@ -99,7 +99,8 @@ const int kOpenExternalLinksInNewTab = 1;
   }
   else if (sender == mSingleWindowMode) {
     [sender setAllowsMixedState:NO];
-    int newState = ([sender state] == NSOnState) ? nsIBrowserDOMWindow::OPEN_NEWTAB : nsIBrowserDOMWindow::OPEN_NEWWINDOW;
+    // Cast to avoid "enumeral mismatch" warning - thanks a lot, xpidl.
+    int newState = ([sender state] == NSOnState) ? (int)nsIBrowserDOMWindow::OPEN_NEWTAB : (int)nsIBrowserDOMWindow::OPEN_NEWWINDOW;
     [self setPref:"browser.link.open_newwindow" toInt:newState];
   }
